Move do_not_recommend to diagnostic folder
