+++ compiler/rustc_attr_parsing/src/attributes/diagnostic/do_not_recommend.rs
