+++ compiler/rustc_attr_parsing/src/attributes/mod.rs
@@ -40,7 +40,6 @@ pub(crate) mod crate_level;
 pub(crate) mod debugger;
 pub(crate) mod deprecation;
 pub(crate) mod diagnostic;
-pub(crate) mod do_not_recommend;
 pub(crate) mod doc;
 pub(crate) mod dummy;
 pub(crate) mod inline;
